Fix: IRepository Get using wrong type

+++ src/TicketSales.Core/TicketSales.Core.Application/Contracts/Persistence/IRepository.cs
@@ -7,7 +7,7 @@ namespace TicketSales.Core.Application.Contracts.Persistence
     public interface IRepository<T> where T : BaseEntity
     {
         Task<IReadOnlyList<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(T entity);
